Рефакторинг фабрики текстур и прочего
