+++ src/main/java/amaralus/apps/hackandslash/graphics/gpu/texture/TextureFactory.java
@@ -1,93 +1,62 @@
 package amaralus.apps.hackandslash.graphics.gpu.texture;
 
-import amaralus.apps.hackandslash.common.ValueEnum;
+import amaralus.apps.hackandslash.io.FileLoadService;
 import amaralus.apps.hackandslash.io.data.ImageData;
+import amaralus.apps.hackandslash.resources.ResourceManager;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-import java.nio.ByteBuffer;
-import java.util.HashMap;
-import java.util.Map;
-
-import static org.lwjgl.opengl.GL11.*;
-import static org.lwjgl.opengl.GL30.glGenerateMipmap;
+import static amaralus.apps.hackandslash.graphics.gpu.texture.PixelFormat.RED;
+import static amaralus.apps.hackandslash.graphics.gpu.texture.PixelFormat.RGBA;
+import static amaralus.apps.hackandslash.graphics.gpu.texture.TextureFilter.*;
+import static amaralus.apps.hackandslash.graphics.gpu.texture.TextureParameterName.*;
+import static amaralus.apps.hackandslash.graphics.gpu.texture.TextureWrapMode.REPEAT;
 
 @Component
+@RequiredArgsConstructor
+@Slf4j
 public class TextureFactory {
 
-    public TextureFactoryBuilder newTexture(String resourceName) {
-        return new TextureFactoryBuilder(resourceName);
+    private final ResourceManager resourceManager;
+    private final FileLoadService fileLoadService;
+
+    public void produceSpriteTexture(String textureName) {
+        logTextureLoad(textureName);
+
+        var imageData = fileLoadService.loadImageData("sprites/" + textureName + ".png");
+
+        var texture = new TextureBuilder(textureName)
+                .imageData(imageData)
+                .pixelFormat(RGBA)
+                .generateMipmap()
+                .param(WRAP_S, REPEAT)
+                .param(WRAP_T, REPEAT)
+                .param(MIN_FILTER, NEAREST)
+                .param(MAG_FILTER, NEAREST)
+                .produce();
+
+        resourceManager.addResource(texture);
     }
 
-    public static class TextureFactoryBuilder {
+    public Texture produceFontTexture(String textureName, ImageData imageData) {
+        logTextureLoad(textureName);
 
-        private final Map<TextureParameterName, ValueEnum> params = new HashMap<>();
-        private final String resourceName;
+        var texture = new TextureBuilder(textureName)
+                .imageData(imageData)
+                .pixelFormat(RED)
+                .generateMipmap()
+                .param(WRAP_S, CLAMP_TO_EDGE)
+                .param(WRAP_T, CLAMP_TO_EDGE)
+                .param(MIN_FILTER, LINEAR)
+                .param(MAG_FILTER, LINEAR)
+                .produce();
 
-        private int width;
-        private int height;
-        private ByteBuffer pixels;
-        private PixelFormat pixelFormat;
-        private boolean generateMipmap;
+        resourceManager.addResource(texture);
+        return texture;
+    }
 
-        public TextureFactoryBuilder(String resourceName) {
-            this.resourceName = resourceName;
-        }
-
-        public TextureFactoryBuilder width(int width) {
-            this.width = width;
-            return this;
-        }
-
-        public TextureFactoryBuilder height(int height) {
-            this.height = height;
-            return this;
-        }
-
-        public TextureFactoryBuilder pixels(ByteBuffer pixels) {
-            this.pixels = pixels;
-            return this;
-        }
-
-        public TextureFactoryBuilder imageData(ImageData imageData) {
-            width = imageData.getWidth();
-            height = imageData.getHeight();
-            pixels = imageData.getImageBytes();
-            return this;
-        }
-
-        public TextureFactoryBuilder pixelFormat(PixelFormat pixelFormat) {
-            this.pixelFormat = pixelFormat;
-            return this;
-        }
-
-        public TextureFactoryBuilder generateMipmap() {
-            generateMipmap = true;
-            return this;
-        }
-
-        public TextureFactoryBuilder param(TextureParameterName parameterName, ValueEnum parameterValue) {
-            params.put(parameterName, parameterValue);
-            return this;
-        }
-
-        public Texture produce() {
-            var texture = new Texture(resourceName, glGenTextures(), width, height);
-
-            texture.bind();
-            applyParams();
-
-            var format = pixelFormat.getValue();
-            glTexImage2D(GL_TEXTURE_2D, 0, format, width, height, 0, format, GL_UNSIGNED_BYTE, pixels);
-
-            if (generateMipmap)
-                glGenerateMipmap(GL_TEXTURE_2D);
-
-            texture.unbind();
-            return texture;
-        }
-
-        private void applyParams() {
-            params.forEach((name, value) -> glTexParameteri(GL_TEXTURE_2D, name.getValue(), value.getValue()));
-        }
+    private void logTextureLoad(String textureName) {
+        log.debug("Загрузка текстуры {}", textureName);
     }
 }
